fix : `MessageBusSubscriber` in `NotificationService` cant make connection to RabbitMQ server.
cause : the `MessageBusSubscriber` was establishing a connection to the RabbitMQ server in its constructor. this seems to be too fast, and RabbitMQ had not yet set things up.

solved with : moved the connection logic to the ExecuteAsync method. by which we delayed the connection to the RabbitMQ server until it is fully set up.